requirements.txt

fastapi
uvicorn
requests
beautifulsoup4
trafilatura
psycopg2-binary
gunicorn
APScheduler
